+++ pkg/crypto/puzzle/puzzle_test.go
@@ -1,11 +1,11 @@
 package puzzle
 
 import (
-	"math/rand"
+	"runtime"
 	"testing"
-	"time"
 
 	"github.com/number571/go-peer/pkg/crypto/hashing"
+	"github.com/number571/go-peer/pkg/encoding"
 	testutils "github.com/number571/go-peer/test/_data"
 )
 
@@ -49,69 +49,13 @@ func TestPuzzle(t *testing.T) {
 	}
 }
 
-/*
-BenchmarkPuzzle/20-bit-12                    100         451693442 ns/op
---- BENCH: BenchmarkPuzzle/20-bit-12
-    puzzle_test.go:90: Timer: 1.267312821s
-    puzzle_test.go:90: Timer: 45.169327733s
-BenchmarkPuzzle/21-bit-12                    100         854375738 ns/op
---- BENCH: BenchmarkPuzzle/21-bit-12
-    puzzle_test.go:90: Timer: 1.881050349s
-    puzzle_test.go:90: Timer: 1m25.437556403s
-BenchmarkPuzzle/22-bit-12                    100        1746838647 ns/op
---- BENCH: BenchmarkPuzzle/22-bit-12
-    puzzle_test.go:90: Timer: 7.615185783s
-    puzzle_test.go:90: Timer: 2m54.6838481s
-*/
+func TestMultiPuzzle(t *testing.T) {
+	t.Parallel()
 
-// go test -bench=BenchmarkPuzzle -benchtime=100x
-func BenchmarkPuzzle(b *testing.B) {
-	benchTable := []struct {
-		name     string
-		function func([]byte, uint64) uint64
-	}{
-		{
-			name:     "20-bit",
-			function: NewPoWPuzzle(20).ProofBytes,
-		},
-		{
-			name:     "21-bit",
-			function: NewPoWPuzzle(21).ProofBytes,
-		},
-		{
-			name:     "22-bit",
-			function: NewPoWPuzzle(22).ProofBytes,
-		},
+	parallel := uint64(runtime.GOMAXPROCS(0))
+	puzzle := NewPoWPuzzle(4)
+	for i := uint64(0); i < 1_000; i++ {
+		arr := encoding.Uint64ToBytes(i)
+		_ = puzzle.ProofBytes(arr[:], parallel)
 	}
-
-	b.ResetTimer()
-
-	for _, t := range benchTable {
-		t := t
-		b.Run(t.name, func(b *testing.B) {
-			b.StopTimer()
-			randomBytes := make([][]byte, 0, b.N)
-			for i := 0; i < b.N; i++ {
-				randomBytes = append(randomBytes, testPseudoRandomBytes(i))
-			}
-			b.StartTimer()
-
-			now := time.Now()
-			for i := 0; i < b.N; i++ {
-				_ = t.function(randomBytes[i], 1)
-			}
-			end := time.Since(now)
-
-			b.Log("Timer:", end)
-		})
-	}
-}
-
-func testPseudoRandomBytes(pSeed int) []byte {
-	r := rand.New(rand.NewSource(int64(pSeed)))
-	result := make([]byte, 0, 16)
-	for i := 0; i < 16; i++ {
-		result = append(result, byte(r.Intn(256)))
-	}
-	return result
 }